feat(eckhart): implement BootloaderUI
- implementation using the new bootloader components
- the implementation sits in the separate file `ui_bootloader.rs`
